Fix casting function (fixed issue #70)

+++ code/Helper/Data.php
@@ -386,7 +386,7 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
 
     private function try_cast($value)
     {
-        if (is_numeric($value) && intval($value) == intval(floatval($value)))
+        if (is_numeric($value) && floatval($value) == floatval(intval($value)))
             return intval($value);
 
         if (is_numeric($value))
@@ -494,7 +494,6 @@ class Algolia_Algoliasearch_Helper_Data extends Mage_Core_Helper_Abstract
         }
 
         $customData = array_merge($customData, $defaultData);
-
 
         if (isset($customData['price'])) {
             $customData['price'] = floatval($customData['price']);
